requirements.txt
absl-py
annotated-types
anyio
apeye
apeye-core
astunparse
certifi
charset-normalizer
click
dnspython
domdf-python-tools
email_validator
exceptiongroup
fastapi
fastapi-cli
flatbuffers
gast
google-pasta
grpcio
gunicorn
h11
h5py
httpcore
httptools
httpx
idna
importlib_metadata
Jinja2
joblib
keras
libclang
Markdown
markdown-it-py
MarkupSafe
mdurl
ml-dtypes
namex
natsort
numpy
opt-einsum
optree
orjson
packaging
pillow
platformdirs
protobuf
pydantic
pydantic_core
Pygments
pypi-json
python-dotenv
python-multipart
PyYAML
requests
rich
scikit-learn
scipy
shellingham
six
sniffio
starlette
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
typer
typing_extensions
ujson
urllib3
uvicorn
uvloop
watchfiles
websockets
Werkzeug
wrapt
zipp
